move workflow tests to own file
