+++ corehq/apps/app_manager/tests/__init__.py
@@ -27,6 +27,7 @@ try:
     from .test_translations import *
     from corehq.apps.app_manager.tests.test_case_meta import *
     from corehq.apps.app_manager.tests.test_case_list_form import *
+    from corehq.apps.app_manager.tests.test_form_workflow import *
 except ImportError, e:
     # for some reason the test harness squashes these so log them here for clarity
     # otherwise debugging is a pain
